added contract address in solution

+++ contracts/ERC918.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.5.1;
 
 interface ERC918 {
-    function mint(uint256 nonce, uint256 targetDifficulty) external returns (bool success);
+    function mint(uint nonce, uint targetDifficulty) external returns (bool success);
 
     function getChallengeNumber() external view returns (bytes32);
 
-    function getMiningDifficulty() external view returns (uint256);
+    function getMiningDifficulty() external view returns (uint);
 
-    function getMiningTarget() external view returns (uint256);
+    function getMiningTarget() external view returns (uint);
 
-    function getMiningReward() external view returns (uint256);
+    function getMiningReward() external view returns (uint);
 }
